.templates/python/app/requirements.txt
paho-mqtt==1.6.1
pywatchman==1.4.1
